// models/sales_folding.model.ts
export type SalesFoldingState = {
  data: {
    sals_fold_id: string;
    sals_fold_date: string;
    sals_fold_company_name: string;
    sals_fold_tel: string;
    sals_fold_email: string;
    sals_fold_doc_type: string;
    sals_fold_printing_type: string;
    sals_fold_amount: string;
    sals_fold_quotation_request: string;
    sals_fold_finished_size: string;
    sals_fold_open_size: string;
    sals_fold_column: string;
    sals_fold_page: string;
    sals_fold_text_paper: string;
    sals_fold_printing: string;
    sals_fold_text_coating: string;
    sals_fold_printing_volume: string;
  };
};